Circular Scroll: Fix scale of speed slider.
The circular scroll delta is measured in radians but the slider for this
parameter had a range of 30 to 300 which is way out. This changes the
slider values to represent degrees between scroll events (in the range 1
to 60), and converts into radians before storing in config or setting
the touchpad parameter.